+++ src/app/code/community/Shopgate/Framework/Model/Cataloginventory/Bubble/Abstract.php
@@ -1,5 +1,4 @@
 <?php
-
 /**
  * Copyright Shopgate Inc.
  *
@@ -19,21 +18,10 @@
  * @copyright Shopgate Inc
  * @license   http://www.apache.org/licenses/LICENSE-2.0 Apache License, Version 2.0
  */
-class Shopgate_Framework_Model_Cataloginventory_Stock extends Shopgate_Framework_Model_Cataloginventory_Stock_Abstract
-{
-    /**
-     * @param Varien_Object $item
-     * @return $this|Mage_CatalogInventory_Model_Stock
-     */
-    public function registerItemSale(Varien_Object $item)
-    {
-        if (Mage::helper("shopgate")->isShopgateApiRequest()
-            && Mage::helper("shopgate/config")->getIsMagentoVersionLower15()
-        ) {
-            return $this;
-        }
 
-        return parent::registerItemSale($item);
-    }
-
-}
+/**
+ * Provide compatibility with Bubble_StockMovements in case it's installed and active
+ *
+ * Class Shopgate_Framework_Model_Cataloginventory_Bubble_Abstract
+ */
+class Shopgate_Framework_Model_Cataloginventory_Bubble_Abstract extends Bubble_StockMovements_Model_CatalogInventory_Stock {}