+++ src/app/code/community/Shopgate/Framework/Model/Observer/Stock.php
@@ -19,21 +19,18 @@
  * @copyright Shopgate Inc
  * @license   http://www.apache.org/licenses/LICENSE-2.0 Apache License, Version 2.0
  */
-class Shopgate_Framework_Model_Cataloginventory_Stock extends Shopgate_Framework_Model_Cataloginventory_Stock_Abstract
+class Shopgate_Framework_Model_Observer_Stock
 {
     /**
-     * @param Varien_Object $item
-     * @return $this|Mage_CatalogInventory_Model_Stock
+     * Check if Bubble_StockMovements is installed and active
      */
-    public function registerItemSale(Varien_Object $item)
+    public function execute()
     {
-        if (Mage::helper("shopgate")->isShopgateApiRequest()
-            && Mage::helper("shopgate/config")->getIsMagentoVersionLower15()
-        ) {
-            return $this;
+        if (Mage::getConfig()->getModuleConfig('Bubble_StockMovements')->is('active', 'true')) {
+            Mage::getConfig()->setNode(
+                'global/models/shopgate/rewrite/cataloginventory/stock/abstract',
+                'Shopgate_Framework_Model_Cataloginventory_Bubble_Abstract'
+            );
         }
-
-        return parent::registerItemSale($item);
     }
-
 }