+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Fixed
+- Compatibility with Magento Marketplace
+
 ## [2.9.68] - 2018-05-16
 ### Fixed
 - Missing shipping methods in cart validation for configurable products